+++ Sources/SwifterSwift/SwiftStdlib/MutableCollectionExtensions.swift
@@ -22,3 +22,19 @@ public extension MutableCollection where Self: RandomAccessCollection {
         sort { $0[keyPath: keyPath] < $1[keyPath: keyPath] }
     }
 }
+
+public extension MutableCollection {
+    /// SwifterSwift: Assign a given value to a field `keyPath` of all elements in the collection.
+    ///
+    /// - Parameter value: The new value of the field
+    /// - Parameter keyPath: The actual field of the element
+    mutating func assignToAll<Value>(value: Value, by keyPath: WritableKeyPath<Element, Value>) {
+        guard !isEmpty else { return }
+
+        var idx = startIndex
+        while idx != endIndex {
+            self[idx][keyPath: keyPath] = value
+            idx = index(after: idx)
+        }
+    }
+}