+++ Tests/SwiftStdlibTests/ArrayExtensionsTests.swift
@@ -8,22 +8,6 @@
 import XCTest
 @testable import SwifterSwift
 
-private struct Person: Equatable {
-    var name: String
-    var age: Int?
-    var location: Location?
-
-    init(name: String, age: Int?, location: Location? = nil) {
-        self.name = name
-        self.age = age
-        self.location = location
-    }
-}
-
-private struct Location: Equatable {
-    let city: String
-}
-
 final class ArrayExtensionsTests: XCTestCase {
 
     func testPrepend() {
@@ -50,6 +34,21 @@ final class ArrayExtensionsTests: XCTestCase {
         var swappedEmptyArray = emptyArray
         swappedEmptyArray.safeSwap(from: 1, to: 3)
         XCTAssertEqual(swappedEmptyArray, emptyArray)
+    }
+
+    func testSortedLike() {
+        let order1 = [1, 2, 3, 4, 5]
+        let candidate1 = [2, 5, 3, 1, 4]
+        XCTAssertEqual(candidate1.sorted(like: order1, keyPath: \.self), order1)
+
+        let candidate2 = [2, 5, 3, 6, 1, 4]
+        XCTAssertEqual(candidate2.sorted(like: order1, keyPath: \.self), [1, 2, 3, 4, 5, 6])
+
+        // swiftlint:disable:next nesting
+        struct TestStruct { let prop: String }
+        let order3 = ["1", "2", "3", "4", "5"]
+        let candidate3 = [TestStruct(prop: "3"), TestStruct(prop: "2"), TestStruct(prop: "2"), TestStruct(prop: "1"), TestStruct(prop: "3")]
+        XCTAssertEqual(candidate3.sorted(like: order3, keyPath: \.prop).map { $0.prop }, ["1", "2", "2", "3", "3"])
     }
 
     func testRemoveAll() {