+++ Tests/UIKitTests/UIImageExtensionsTests.swift
@@ -132,6 +132,36 @@ final class UIImageExtensionsTests: XCTestCase {
         XCTAssertEqual(image?.size, size)
 
         XCTAssertEqual(image?.bytesSize, 787)
+
+        let scale = CGFloat(5.0)
+        let scaledSize = CGSize(width: size.width / scale, height: size.height / scale)
+
+        let scaledImage = UIImage(base64String: base64String, scale: scale)
+        XCTAssertEqual(scaledImage?.size, scaledSize)
+    }
+
+    func testURL() {
+        let bundle = Bundle.init(for: UIImageExtensionsTests.self)
+        guard let swifterSwiftLogo = bundle.url(forResource: "TestImage", withExtension: "png") else { XCTAssert(false, "Swifter Swift Test Image not available, or url is no longer valid."); return}
+        let image = try? UIImage(url: swifterSwiftLogo)
+        XCTAssertNotNil(image)
+
+        let size = CGSize(width: 1000, height: 232)
+        XCTAssertEqual(image?.size, size)
+
+        let scale: CGFloat = 5.0
+        let scaledSize = CGSize(width: size.width / scale, height: size.height / scale)
+
+        let scaledImage = try? UIImage(url: swifterSwiftLogo, scale: scale)
+        XCTAssertNotNil(scaledImage)
+        XCTAssertEqual(scaledImage?.size, scaledSize)
+
+        guard let throwingURL = URL(string: "SwifterSwift://fakeurl/image1") else {
+            XCTAssert(false, "Fake URL cannot be made")
+            return
+        }
+
+        XCTAssertThrowsError(try UIImage(url: throwingURL))
     }
 
     func testTinted() {
@@ -156,6 +186,16 @@ final class UIImageExtensionsTests: XCTestCase {
         XCTAssertNotNil(image.withRoundedCorners(radius: 5))
         XCTAssertNotNil(image.withRoundedCorners(radius: -10))
         XCTAssertNotNil(image.withRoundedCorners(radius: 30))
+    }
+
+    func testPNGBase64String() {
+        let image = UIImage(color: .blue, size: CGSize(width: 1, height: 1))
+        XCTAssertEqual(image.pngBase64String(), "iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAAAXNSR0IArs4c6QAAAERlWElmTU0AKgAAAAgAAYdpAAQAAAABAAAAGgAAAAAAA6ABAAMAAAABAAEAAKACAAQAAAABAAAAAaADAAQAAAABAAAAAQAAAAD5Ip3+AAAADUlEQVQIHWNgYPj/HwADAgH/p+FUpQAAAABJRU5ErkJggg==")
+    }
+
+    func testJPEGBase64String() {
+        let image = UIImage(color: .blue, size: CGSize(width: 1, height: 1))
+        XCTAssertEqual(image.jpegBase64String(compressionQuality: 1), "/9j/4AAQSkZJRgABAQAASABIAAD/4QBYRXhpZgAATU0AKgAAAAgAAgESAAMAAAABAAEAAIdpAAQAAAABAAAAJgAAAAAAA6ABAAMAAAABAAEAAKACAAQAAAABAAAAAaADAAQAAAABAAAAAQAAAAD/7QA4UGhvdG9zaG9wIDMuMAA4QklNBAQAAAAAAAA4QklNBCUAAAAAABDUHYzZjwCyBOmACZjs+EJ+/8AAEQgAAQABAwERAAIRAQMRAf/EAB8AAAEFAQEBAQEBAAAAAAAAAAABAgMEBQYHCAkKC//EALUQAAIBAwMCBAMFBQQEAAABfQECAwAEEQUSITFBBhNRYQcicRQygZGhCCNCscEVUtHwJDNicoIJChYXGBkaJSYnKCkqNDU2Nzg5OkNERUZHSElKU1RVVldYWVpjZGVmZ2hpanN0dXZ3eHl6g4SFhoeIiYqSk5SVlpeYmZqio6Slpqeoqaqys7S1tre4ubrCw8TFxsfIycrS09TV1tfY2drh4uPk5ebn6Onq8fLz9PX29/j5+v/EAB8BAAMBAQEBAQEBAQEAAAAAAAABAgMEBQYHCAkKC//EALURAAIBAgQEAwQHBQQEAAECdwABAgMRBAUhMQYSQVEHYXETIjKBCBRCkaGxwQkjM1LwFWJy0QoWJDThJfEXGBkaJicoKSo1Njc4OTpDREVGR0hJSlNUVVZXWFlaY2RlZmdoaWpzdHV2d3h5eoKDhIWGh4iJipKTlJWWl5iZmqKjpKWmp6ipqrKztLW2t7i5usLDxMXGx8jJytLT1NXW19jZ2uLj5OXm5+jp6vLz9PX29/j5+v/bAEMAAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAf/bAEMBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAf/dAAQAAf/aAAwDAQACEQMRAD8A/jnr/v4P5XP/2Q==")
     }
 
 }